Autocommit Wed Nov 20 22:14:42 NZDT 2013
Writers who use a computer swear to its liberating power in tones that bear
witness to the apocalyptic power of a new divinity.  Their conviction results
from something deeper than mere gratitude for the computer's conveniences.
Every new medium of writing brings about new intensities of religious belief
and new schisms among believers.  In the 16th century the printed book helped
make possible the split between Catholics and Protestants.  In the 20th
century this history of tragedy and triumph is repeating itself as a farce.
Those who worship the Apple computer and those who put their faith in the IBM
PC are equally convinced that the other camp is damned or deluded.  Each cult
holds in contempt the rituals and the laws of the other.  Each thinks that it
is itself the one hope for salvation.
		-- Edward Mendelson, 'The New Republic', February 22, 1988