Autocommit Thu Nov 21 08:36:43 NZDT 2013
Speaking of love, one problem that recurs more and more frequently these
days, in books and plays and movies, is the inability of people to communicate
with the people they love; Husbands and wives who can't communicate, children
who can't communicate with their parents, and so on.  And the characters in
these books and plays and so on (and in real life, I might add) spend hours
bemoaning the fact that they can't communicate.  I feel that if a person can't
communicate, the very _____least he can do is to shut up!
		-- Tom Lehrer, 'That Was the Year that Was'