Autocommit Thu Nov 21 20:47:14 NZDT 2013
...cyberpunk wants to see the mind as mechanistic & duplicable,
challenging basic assumptions about the nature of individuality & self.
That seems all the better reason to assume that cyberpunk art & music is
essentially mindless garbagio. Willy certainly addressed this idea in
'Count Zero,' with Katatonenkunst, the automatic box-maker and the girl's
observation that the real art was the building of the machine itself,
rather than its output.
-- Eliot Handelman